REF: Camera ready CNN pipeline
