+++ baselines/cnn/1D_CNN_Linear_probing.py
@@ -1,136 +1,18 @@
 import argparse
+import sys
 
 import numpy as np
 import pandas as pd
-import sklearn
-import sklearn.metrics
 import torch
-from torch import nn
+import wandb
 from torch.utils.data import DataLoader
 
-import wandb
-
-
-def evaluate(model, dataloader, device):
-    model.eval()
-
-    y_true_all = []
-    y_pred_all = []
-
-    for sequences, y_true in dataloader:
-        sequences = sequences.to(device)
-        y_true = y_true.to(device)
-
-        with torch.no_grad():
-            logits = model(sequences)
-            y_pred = torch.argmax(logits, dim=-1)
-
-        y_true_all.append(y_true.cpu().numpy())
-        y_pred_all.append(y_pred.cpu().numpy())
-
-    # Concatenate the targets and predictions from each batch
-    y_true = np.concatenate(y_true_all)
-    y_pred = np.concatenate(y_pred_all)
-
-    print(f"y_true: {y_true.shape}, y_pred: {y_pred.shape}")
-
-    for x in np.unique(y_pred):
-        if x not in y_true:
-            print(f"Predicted label {x} is not in the GT")
-
-    # Create results dictionary
-    results = {}
-    results["count"] = len(y_true)
-    # Note that these evaluation metrics have all been converted to percentages
-    results["accuracy"] = 100.0 * sklearn.metrics.accuracy_score(y_true, y_pred)
-    results["accuracy-balanced"] = 100.0 * sklearn.metrics.balanced_accuracy_score(y_true, y_pred)
-    results["f1-micro"] = 100.0 * sklearn.metrics.f1_score(y_true, y_pred, average="micro")
-    results["f1-macro"] = 100.0 * sklearn.metrics.f1_score(y_true, y_pred, average="macro")
-    results["f1-support"] = 100.0 * sklearn.metrics.f1_score(y_true, y_pred, average="weighted")
-
-    print(" Evaluation results:")
-    for k, v in results.items():
-        if k == "count":
-            print(f"  {k + ' ':.<21s}{v:7d}")
-        elif k in ["max_ram_mb", "peak_vram_mb"]:
-            print(f"  {k + ' ':.<24s} {v:6.2f} MB")
-        else:
-            print(f"  {k + ' ':.<24s} {v:6.2f} %")
-
-    wandb.log({f"eval/{k}": v for k, v in results.items()})
-
-
-# CNN model architecture
-class CNNModel(nn.Module):
-    def __init__(self, n_input, n_output):
-        super(CNNModel, self).__init__()
-        self.conv1 = nn.Conv2d(n_input, 64, kernel_size=(3, 3), stride=1, padding=1)
-        self.bn1 = nn.BatchNorm2d(64)
-        self.maxpool1 = nn.MaxPool2d(kernel_size=(3, 1))
-        self.conv2 = nn.Conv2d(64, 32, kernel_size=(3, 3), stride=1, padding=1)
-        self.bn2 = nn.BatchNorm2d(32)
-        self.maxpool2 = nn.MaxPool2d(kernel_size=(3, 1))
-        self.conv3 = nn.Conv2d(32, 16, kernel_size=(3, 3), stride=1, padding=1)
-        self.bn3 = nn.BatchNorm2d(16)
-        self.maxpool3 = nn.MaxPool2d(kernel_size=(3, 1))
-        self.flatten = nn.Flatten()
-        self.bn4 = nn.BatchNorm1d(1920)
-        self.dense1 = nn.Linear(1920, 500)
-        self.dense2 = nn.Linear(500, n_output)
-
-    def forward(self, x):
-        x = self.conv1(x)
-        x = nn.ReLU()(x)
-        x = self.bn1(x)
-        x = self.maxpool1(x)
-        x = self.conv2(x)
-        x = nn.ReLU()(x)
-        x = self.bn2(x)
-        x = self.maxpool2(x)
-        x = self.conv3(x)
-        x = nn.ReLU()(x)
-        x = self.bn3(x)
-        x = self.maxpool3(x)
-        x = self.flatten(x)
-        x = self.bn4(x)
-        x = self.dense1(x)
-        out = nn.Tanh()(x)
-        out = self.dense2(out)
-        return out, x
-
-
-def data_from_df(df, target_level, label_pipeline):
-    barcodes = df["nucleotides"].to_list()
-    species = df[target_level].to_list()
-
-    species = np.array(list(map(label_pipeline, species)))
-
-    print(f"[INFO]: There are {len(barcodes)} barcodes")
-    # Number of training samples and entire data
-    N = len(barcodes)
-
-    # Reading barcodes and labels into python list
-    labels = []
-
-    for i in range(N):
-        if len(barcodes[i]) > 0:
-            barcodes.append(barcodes[i])
-            labels.append(species[i])
-
-    sl = 660  # Max_length
-
-    nucleotide_dict = {"A": 0, "C": 1, "G": 2, "T": 3, "N": 4}
-
-    X = np.zeros((N, sl, 5), dtype=np.float32)
-    for i in range(N):
-
-        for j in range(sl):
-            if len(barcodes[i]) > j:
-                k = nucleotide_dict[barcodes[i][j]]
-                X[i][j][k] = 1.0
-
-    # print(X.shape, )
-    return X, np.array(labels)
+sys.path.append(".")
+from baselines.cnn.cnn_utils import (
+    CNNModel,
+    data_from_df,
+    evaluate,
+)
 
 
 def run(config):
